Cleanup OFMessageFutures
Removed stale code for OFMessageFuture and OFStatisticsFuture.
Change the way we keep track of pending Futures: instead of registering
an IOFSwitchListener we now cancel all pending futures when a switch is
removed from the active list.  In the future we might want to allow slave
controllers to send stats request then we need to move the cancel to
switch disconnect.

[#26756797]