Add support for BSN_L2_TABLE_SET to the state machine.
Still work in progress but pushing to unrelated unit test problems can
be solved.

Moved the core and access priorities for flow mods from Forwarding to
Controller and make it a property of IOFSwitch.

Waiting for L2_SET_TABLE_REPLY message to be speced.

Need to add unit tests for L2_TABLE_SET